get-started-python/requirements.txt
Flask==0.12.2
cf-deployment-tracker==1.0.4
cloudant==2.4.0
nltk==3.2.2
pypdf==1.13
pdf==1.0
PyPDF2==1.26.0
python-docx==0.8.6
pytesseract==0.2.0
pillow==5.1.0
